Adding Query on List

+++ rasa/actions/actions.py
@@ -14,6 +14,17 @@ from rasa_sdk.events import SlotSet
 import rasa_sdk.events
 import smtplib
 import os
+from typing import Text
+from pymongo import MongoClient
+import re
+import pprint
+from bson.objectid import ObjectId     
+from collections import OrderedDict
+import json
+import itertools
+import colorama
+
+from colorama import Fore, Back, Style
 from pymongo import MongoClient
 
 email=""
@@ -69,25 +80,8 @@ class ActionShowWheater(Action):
         temp_city = ((api_data['main']['temp'])-273.15)
         weather_desc = api_data['weather'][0]['description']
         print(tracker.slots['city'])
-        dispatcher.utter_message(text=f"{'The weataomher in ' + city + ' is '+ weather_desc + ' and the temperature is '+ str(round(temp_city,1)) +'°'}")
+        dispatcher.utter_message(text=f"{'The weather in ' + city + ' is '+ weather_desc + ' and the temperature is '+ str(round(temp_city,1)) +'°'}")
     
-        client = MongoClient('localhost:27017')
-        mydb= client['prova']
-        mycol= mydb['characters']
-        print(mycol)
-
-        data={'name':'oca toro'}
-        mycol.insert_one(data)
-
-        print(client.list_database_names())
-
-        print(mydb.list_collection_names())
-
-        for x in mycol.find({"name":"pino peppe"}) :
-            print(x) 
-
-        mycol.delete_one({"name":"pino peppe"})
-
         return []
 
 
@@ -116,6 +110,8 @@ class ActionAddTodo(Action):
 
 
 
+
+
 class ActionCompleteTodo(Action):
 
     def name(self) -> Text:
@@ -127,16 +123,74 @@ class ActionCompleteTodo(Action):
         entities = tracker.latest_message['entities']
         for e in entities:
             task=e['value']
-        print(task)
-        dispatcher.utter_message(text=f"{'You have completed your task: ' + task }")
                 
         client = MongoClient('localhost:27017')
         mydb= client['prova']
         mycol= mydb['characters']
-        print(mycol)
+        
+        new_list=[]
+        id_list=[]
+        
+        data = {'name' : task}
+        query = mycol.find({ 'name': task })
+        for i in query:
+            print(i)
+        check = format(query.retrieved)
+        print(check)
+        if int(check) == 1:  
+            query = mycol.delete_one(data)
+            dispatcher.utter_message(text=f"{'You have completed your task: ' + task }")
+        else:
+            li = list(task.split(" "))
+            for L in range(1, len(li)+1):
+                for subset in itertools.combinations(li, L):
+                    cont = L
+                    string = ' { "$and" : [ { "name": { "$regex" : ".*'+ subset[0] +'.*" } } ] }'
+                    for X in range(1, L):
+                        string = string.replace('] }',', { "name": { "$regex" : ".*'+ subset[X] +'.*" } } ] }')
 
-        data={'name':task}
-        mycol.delete_one(data)
+            data = json.loads(string)
+
+            result= mycol.find( data )
+            for x in result:
+                id=x['_id']
+                id_list=[]
+                
+                id_list.append(str(L))
+                id_list.append(str(id))
+                new_list.append(id_list)
+
+            res = list(set(tuple(sorted(sub)) for sub in new_list))
+            if len(res) != 0:
+                max_value = max(res)
+            else: max_value=0
+
+            for i in res:
+                for j in res:
+                    if i[1]==j[1] and i[0]!=j[0] and i[0]<j[0]:
+                        try:
+                            res.remove(i)
+                        except:
+                            pass
+            
+            new_res=[]
+            for el in res:
+                if int(el[0]) >= int(max_value[0]): 
+                    new_res.append(el)
+            
+            if len(new_res)==0:
+                dispatcher.utter_message(text=f"{'I am sorry, no task found :( ' }")
+            elif len(new_res)==1:
+                query = mycol.delete_one(data)
+                dispatcher.utter_message(text=f"{'You have completed your task: ' + task }")
+            else: 
+                mess=""
+                for i in range(0,len(new_res)):
+                    for j in mycol.find({"_id": ObjectId(new_res[i][1])}):
+                        mess = mess + j['name'] + '\n'
+                mess= '\n' + mess
+                dispatcher.utter_message(text=f"{'I found those task: ' + mess }")
+
 
 
 
@@ -165,7 +219,7 @@ class ActionAskTodo(Action):
             single_task = '\nNumber ' + str(i) + ': ' + x['name'] + '.'
             list = list + single_task
         
-        dispatcher.utter_message(text=f"{'You have ' + str(num_task) + ' in your TODO list: ' + list }")
+        dispatcher.utter_message(text=f"{'You have ' + str(num_task) + ' tasks in your TODO list: ' + list }")
